+++ src/app/commands/bounty/Bounty.ts
@@ -3,7 +3,7 @@ import create from '../../service/bounty/create/new';
 import list from '../../service/bounty/list';
 import claim from '../../service/bounty/claim';
 import validate from '../../service/bounty/create/validate';
-import serviceUtils from '../../service/ServiceUtils';
+import serviceUtils from '../../utils/ServiceUtils';
 import constants from '../../constants';
 import { GuildMember } from 'discord.js';
